+++ app/controllers/documents_controller.rb
@@ -1,10 +1,19 @@
 class DocumentsController < ApplicationController
   before_action :authenticate_user!
+  before_action :set_viewable_project, only: [ :file ]
   before_action :set_editable_project, only: [ :index, :show, :new, :edit, :create, :update, :destroy ]
-  before_action :redirect_without_project, only: [ :index, :show, :new, :edit, :create, :update, :destroy ]
+  before_action :redirect_without_project, only: [ :index, :show, :new, :edit, :create, :update, :destroy, :file ]
+  before_action :set_viewable_document, only: [ :file ]
   before_action :set_editable_document, only: [ :show, :edit, :update, :destroy ]
-  before_action :redirect_without_document, only: [ :show, :edit, :update, :destroy ]
+  before_action :redirect_without_document, only: [ :show, :edit, :update, :destroy, :file ]
 
+  def file
+    if @document.file.size > 0
+      send_file File.join( CarrierWave::Uploader::Base.root, @document.file.url )
+    else
+      render nothing: true
+    end
+  end
 
   # GET /documents
   # GET /documents.json
@@ -71,6 +80,10 @@ class DocumentsController < ApplicationController
   end
 
   private
+
+    def set_viewable_document
+      @document = @project.documents.find_by_id(params[:id])
+    end
 
     def set_editable_document
       @document = @project.documents.find_by_id(params[:id])